// src/types/types.ts
export interface Review {
  id: number; // Review ID
  store: number; // Store ID
  user: {
    username: string; // User's username
    profile_image: string; // User's profile image URL
  }; // Nested user object
  rating: number; // Rating given in the review
  comment: string; // Review comment
  createdAt: string; // Date and time the review was created
}

export interface ReviewsProps {
  reviews: Review[];
  storeId: number;
}

// Updated ReviewState Interface
export interface ReviewState {
  reviews: Review[]; // Array of reviews
  loading: boolean; // Indicates whether the reviews are being fetched
  error: string | null; // Holds any error message, or null if no error
}